+++ src/components/Form/index.tsx
@@ -1,0 +1,4 @@
+import FormInput from './FormInput';
+import FormButton from './FormButton';
+
+export { FormButton, FormInput };
