+++ src/pages/Login/components/EmailLoginForm.tsx
@@ -1,7 +1,6 @@
 import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
-import FormInput from '@/components/Form/FormInput';
-import FormButton from '@/components/Form/FormButton';
+import { FormInput, FormButton } from '@/components/Form';
 import { ILoginForm } from '../types';
 import { LoginForm } from './EmailLoginForm.style';
